refactor(tests)!: reorganize test suite and fix parsing bugs
BREAKING CHANGE: Test files have been reorganized with new naming
convention

- Reorganize tests into logical groups by functionality:
  - test_parse_basic.py: elements, attributes, nesting, unicode
  - test_parse_parameters.py: force_cdata, attr_prefix, cdata_key, etc.
  - test_parse_lists.py: force_list functionality
  - test_parse_namespaces.py: namespace handling
  - test_parse_callbacks.py: postprocessor
  - test_parse_special.py: CDATA, comments, XML declarations
  - test_parse_input.py: input sources (str, bytes, file-like,
    generators)
  - test_unparse.py: dict to XML conversion
  - test_roundtrip.py: parse/unparse round-trip tests

- Fix namespace handling bugs:
  - Add @xmlns when namespaced attributes are used without mapping
  - Skip separator when namespace maps to empty string

- Fix XML entity decoding in attributes using unescape_value()
