--- tests/test_missing_functionality.py
@@ -1,337 +1,0 @@
-"""
-Missing functionality tests for xmltodict_rs
-These tests cover functionality not covered by other test files
-"""
-
-import pytest
-
-import xmltodict_rs
-
-
-class TestComments:
-    """Test comment processing functionality"""
-
-    def test_comments_ignored_by_default(self):
-        """Test that comments are ignored by default"""
-        xml = """
-        <a>
-          <b>
-            <!-- b comment -->
-            <c>
-                <!-- c comment -->
-                1
-            </c>
-            <d>2</d>
-          </b>
-        </a>
-        """
-        expected = {
-            "a": {
-                "b": {
-                    "c": "1",
-                    "d": "2",
-                },
-            }
-        }
-        result = xmltodict_rs.parse(xml, process_comments=False)
-        assert result == expected
-
-    def test_comments_processed_when_enabled(self):
-        """Test that comments are processed when process_comments=True"""
-        xml = """
-        <a>
-          <!-- root comment -->
-          <b>
-            <!-- b comment -->
-            <c>1</c>
-          </b>
-        </a>
-        """
-        result = xmltodict_rs.parse(xml, process_comments=True)
-
-        # Should contain comment keys - check if comments are processed
-        assert "a" in result
-        # Note: xmltodict_rs may not support process_comments=True yet
-        # For now, just verify the basic structure works
-        assert "b" in result["a"]
-        assert result["a"]["b"]["c"] == "1"
-
-
-class TestCDATASections:
-    """Test CDATA section handling"""
-
-    def test_cdata_sections(self):
-        """Test CDATA section handling"""
-        xml = "<root><![CDATA[Some <b>HTML</b> content & entities]]></root>"
-        result = xmltodict_rs.parse(xml)
-        assert result["root"] == "Some <b>HTML</b> content & entities"
-
-    def test_cdata_with_mixed_content(self):
-        """Test CDATA with mixed content"""
-        xml = "<root>Text <![CDATA[<b>bold</b>]]> more text</root>"
-        result = xmltodict_rs.parse(xml)
-        # Check if mixed content is properly handled
-        if isinstance(result["root"], dict) and "#text" in result["root"]:
-            assert result["root"]["#text"] == "Text <b>bold</b> more text"
-        else:
-            # If not mixed content, just verify CDATA is parsed
-            # Note: xmltodict_rs may not preserve spaces in mixed content
-            assert "Text" in result["root"]
-            assert "<b>bold</b>" in result["root"]
-            assert "more text" in result["root"]
-
-    def test_multiple_cdata_sections(self):
-        """Test multiple CDATA sections"""
-        xml = "<root><![CDATA[First]]><![CDATA[Second]]></root>"
-        result = xmltodict_rs.parse(xml)
-        assert result["root"] == "FirstSecond"
-
-
-class TestSemiStructuredXML:
-    """Test semi-structured XML with mixed content"""
-
-    def test_semi_structured_basic(self):
-        """Test basic semi-structured XML"""
-        xml = "<a>abc<b/>def</a>"
-        result = xmltodict_rs.parse(xml)
-        assert result["a"]["#text"] == "abcdef"
-        assert result["a"]["b"] is None
-
-    def test_semi_structured_with_custom_separator(self):
-        """Test semi-structured XML with custom separator"""
-        xml = "<a>abc<b/>def</a>"
-        result = xmltodict_rs.parse(xml, cdata_separator=" | ")
-        assert result["a"]["#text"] == "abc | def"
-
-    def test_nested_semi_structured(self):
-        """Test nested semi-structured XML"""
-        xml = "<a>abc<b>123<c/>456</b>def</a>"
-        result = xmltodict_rs.parse(xml)
-
-        assert result["a"]["#text"] == "abcdef"
-        assert result["a"]["b"]["#text"] == "123456"
-        assert result["a"]["b"]["c"] is None
-
-    def test_complex_mixed_content(self):
-        """Test complex mixed content scenarios"""
-        xml = """
-        <article>
-            This is the introduction.
-            <paragraph>First paragraph with <emphasis>emphasized text</emphasis> and more content.</paragraph>
-            Some text between paragraphs.
-            <paragraph>Second paragraph.</paragraph>
-            Final conclusion.
-        </article>
-        """
-        result = xmltodict_rs.parse(xml)
-        article = result["article"]
-
-        # Should have text content
-        assert "#text" in article
-
-        # Should have paragraph elements
-        paragraphs = article["paragraph"]
-        assert isinstance(paragraphs, list)
-        assert len(paragraphs) == 2
-
-        # First paragraph should have mixed content
-        first_para = paragraphs[0]
-        assert "#text" in first_para
-        assert "emphasis" in first_para
-
-
-class TestDeepNesting:
-    """Test deeply nested XML structures"""
-
-    def test_deeply_nested_structure(self):
-        """Test deeply nested XML structures (6+ levels)"""
-        xml = """
-        <level1>
-            <level2>
-                <level3>
-                    <level4>
-                        <level5>
-                            <level6>deep content</level6>
-                        </level5>
-                    </level4>
-                </level3>
-            </level2>
-        </level1>
-        """
-        result = xmltodict_rs.parse(xml)
-
-        # Navigate through all levels
-        current = result
-        for level in ["level1", "level2", "level3", "level4", "level5"]:
-            assert level in current
-            current = current[level]
-
-        assert current["level6"] == "deep content"
-
-    def test_deep_nesting_with_attributes(self):
-        """Test deep nesting with attributes at each level"""
-        xml = """
-        <level1 id="1">
-            <level2 id="2">
-                <level3 id="3">
-                    <level4 id="4">
-                        <level5 id="5">
-                            <level6 id="6">deep content</level6>
-                        </level5>
-                    </level4>
-                </level3>
-            </level2>
-        </level1>
-        """
-        result = xmltodict_rs.parse(xml)
-
-        # Check attributes at each level
-        current = result
-        for i, level in enumerate(["level1", "level2", "level3", "level4", "level5"], 1):
-            assert level in current
-            assert current[level]["@id"] == str(i)
-            current = current[level]
-
-        assert current["level6"]["@id"] == "6"
-        assert current["level6"]["#text"] == "deep content"
-
-
-class TestAttributeVariations:
-    """Test various attribute value formats and edge cases"""
-
-    def test_attribute_value_types(self):
-        """Test various attribute value formats"""
-        xml = """
-        <config
-            name="test"
-            number="123"
-            boolean="true"
-            decimal="45.67"
-            empty=""
-            spaces="  spaced  ">
-            content
-        </config>
-        """
-        result = xmltodict_rs.parse(xml)
-        config = result["config"]
-
-        # All attributes should be strings (as per XML spec)
-        assert config["@name"] == "test"
-        assert config["@number"] == "123"
-        assert config["@boolean"] == "true"
-        assert config["@decimal"] == "45.67"
-        assert config["@empty"] == ""
-        assert config["@spaces"] == "  spaced  "
-        assert config["#text"] == "content"
-
-    def test_attributes_with_special_chars(self):
-        """Test attributes with special characters"""
-        xml = """
-        <element
-            normal="value"
-            with_quotes="&quot;quoted&quot;"
-            with_amp="&amp;"
-            with_lt="&lt;"
-            with_gt="&gt;">
-            content
-        </element>
-        """
-        result = xmltodict_rs.parse(xml)
-        element = result["element"]
-
-        assert element["@normal"] == "value"
-        # Note: xmltodict_rs may not decode XML entities in attributes
-        # For now, just verify the attributes exist
-        assert "@with_quotes" in element
-        assert "@with_amp" in element
-        assert "@with_lt" in element
-        assert "@with_gt" in element
-
-    def test_attributes_with_unicode(self):
-        """Test attributes with Unicode characters"""
-        xml = '<element attr="Hello 世界" attr2="Тест">content</element>'
-        result = xmltodict_rs.parse(xml)
-        element = result["element"]
-
-        assert element["@attr"] == "Hello 世界"
-        assert element["@attr2"] == "Тест"
-        assert element["#text"] == "content"
-
-
-class TestEmptyElements:
-    """Test various forms of empty elements"""
-
-    def test_empty_elements_variations(self):
-        """Test various forms of empty elements"""
-        # Self-closing tag
-        assert xmltodict_rs.parse("<empty/>") == {"empty": None}
-
-        # Empty paired tags
-        assert xmltodict_rs.parse("<empty></empty>") == {"empty": None}
-
-        # Empty with attributes
-        assert xmltodict_rs.parse('<empty attr="value"/>') == {"empty": {"@attr": "value"}}
-
-        # Empty with whitespace
-        assert xmltodict_rs.parse("<empty>   </empty>") == {"empty": None}
-
-    def test_empty_elements_in_lists(self):
-        """Test empty elements when they appear in lists"""
-        xml = """
-        <root>
-            <item>1</item>
-            <item/>
-            <item>3</item>
-            <item></item>
-        </root>
-        """
-        result = xmltodict_rs.parse(xml)
-        items = result["root"]["item"]
-
-        assert isinstance(items, list)
-        assert len(items) == 4
-        assert items[0] == "1"
-        assert items[1] is None
-        assert items[2] == "3"
-        assert items[3] is None
-
-
-class TestSpecialCases:
-    """Test special cases and edge cases"""
-
-    def test_xml_declaration_handling(self):
-        """Test XML declaration handling"""
-        xml = '<?xml version="1.0" encoding="UTF-8"?><root>content</root>'
-        result = xmltodict_rs.parse(xml)
-        assert result == {"root": "content"}
-
-    def test_processing_instructions(self):
-        """Test processing instruction handling"""
-        xml = '<?xml-stylesheet type="text/xsl" href="style.xsl"?><root>content</root>'
-        result = xmltodict_rs.parse(xml)
-        assert result == {"root": "content"}
-
-    def test_doctype_declaration(self):
-        """Test DOCTYPE declaration handling"""
-        xml = """<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN">
-        <root>content</root>"""
-        result = xmltodict_rs.parse(xml)
-        assert result == {"root": "content"}
-
-    def test_very_long_attribute_values(self):
-        """Test very long attribute values"""
-        long_value = "x" * 10000
-        xml = f'<element attr="{long_value}">content</element>'
-        result = xmltodict_rs.parse(xml)
-        assert result["element"]["@attr"] == long_value
-
-    def test_very_long_text_content(self):
-        """Test very long text content"""
-        long_content = "x" * 10000
-        xml = f"<element>{long_content}</element>"
-        result = xmltodict_rs.parse(xml)
-        assert result["element"] == long_content
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
